requirements.txt
# UI
streamlit>=1.35.0

# Core VLM inference
llama-cpp-python>=0.2.90
openai>=1.30.0          # OpenAI-compatible client for llama.cpp server

# Image processing
Pillow>=10.3.0
pdf2image>=1.17.0       # PDF → image (requires poppler)
pypdf>=4.2.0

# Data / schemas
pydantic>=2.7.0
pandas>=2.2.0

# MLOps / logging
mlflow>=2.13.0
python-dotenv>=1.0.1
rich>=13.7.0            # Pretty console output
tqdm>=4.66.0

# Optional: HuggingFace download
huggingface-hub>=0.23.0

# Optional: fine-tuning (heavy — install separately)
# torch>=2.3.0
# transformers>=4.44.0
# peft>=0.11.0
# trl>=0.9.0
# datasets>=2.19.0
# accelerate>=0.30.0
# bitsandbytes>=0.43.0
